chore: rename performance tests
